JEPA_From_Scratch/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0
Pillow>=8.2.0
PyYAML>=6.0
tqdm>=4.61.0
scikit-learn>=0.24.2
matplotlib>=3.4.2
tensorboard>=2.5.0
h5py>=3.2.1
requests>=2.25.1
